Validated pointnet++ with UNET base model
