+++ config.yaml
@@ -6,8 +6,8 @@ data:
 
 training:
     shuffle: True
-    num_workers: 2
-    batch_size: 4
+    num_workers: 6
+    batch_size: 8
     cuda: 1
 
 experiment:
@@ -35,7 +35,7 @@ models:
             convs:
                 input_nc: 3
                 ratios: [0.2, 0.25]
-                radius: [0.1, 0.2]
+                radius: [0.2, 0.4]
                 down_conv_nn: [[3, 64, 64, 128], [131, 128, 128, 256]]
                 up_conv_nn: [[384, 256, 128], [1280, 256, 256]]
                 final_up_conv_nn: [128, 128, 128, 128]
@@ -45,22 +45,3 @@ models:
                 nn: [259, 256, 512, 1024]
             skip: True
             mlp_cls: [128, 128, 128, 128]
-        pointunet:
-            name: pointunet
-            convs:
-                input_nc: 3
-                ratios: [0.2, 0.25]
-                radius: [0.1, 0.2]
-                down_conv_nn: [[3, 64, 64, 128], [131, 128, 128, 256]]
-                up_conv_nn: [[384, 256, 128], [1280, 256, 256]]
-                final_up_conv_nn: [128, 128, 128, 128]
-                up_k: [1, 3, 3]
-            global:
-                aggr: max
-                nn: [259, 256, 512, 1024]
-            skip: True
-            mlp_cls: [128, 128, 128, 128]
-
-             
-                
-
