+++ models/pointnet2/nn.py
@@ -2,91 +2,32 @@ import torch
 from torch import nn
 import torch.nn.functional as F
 from torch.nn import Sequential as Seq, Linear as Lin, ReLU, BatchNorm1d as BN
+from models.base_model import MLP, FPModule, UnetBasedModel
 from torch_geometric.nn import knn_interpolate
 from torch_geometric.nn import PointConv, fps, radius, global_max_pool
+from .modules import SAModule
 
-
-class SAModule(torch.nn.Module):
-    def __init__(self, ratio, r, nn):
-        super(SAModule, self).__init__()
-        self.ratio = ratio
-        self.r = r
-        self.conv = PointConv(local_nn = nn)
-
-    def forward(self, x, pos, batch):
-        idx = fps(pos, batch, ratio=self.ratio)
-        row, col = radius(pos, pos[idx], self.r, batch, batch[idx],
-                          max_num_neighbors=64)
-        edge_index = torch.stack([col, row], dim=0)
-        x = self.conv(x, (pos, pos[idx]), edge_index)
-        pos, batch = pos[idx], batch[idx]
-        return x, pos, batch
-
-
-class GlobalSAModule(torch.nn.Module):
-    def __init__(self, nn):
-        super(GlobalSAModule, self).__init__()
-        self.nn = nn
-
-    def forward(self, x, pos, batch):
-        x = self.nn(torch.cat([x, pos], dim=1))
-        x = global_max_pool(x, batch)
-        pos = pos.new_zeros((x.size(0), 3))
-        batch = torch.arange(x.size(0), device=batch.device)
-        return x, pos, batch
-
-
-def MLP(channels, batch_norm=True):
-    return Seq(*[
-        Seq(Lin(channels[i - 1], channels[i]), ReLU(), BN(channels[i]))
-        for i in range(1, len(channels))
-    ])
-
-
-
-class FPModule(torch.nn.Module):
-    def __init__(self, k, nn):
-        super(FPModule, self).__init__()
-        self.k = k
-        self.nn = nn
-
-    def forward(self, x, pos, batch, x_skip, pos_skip, batch_skip):
-        x = knn_interpolate(x, pos, pos_skip, batch, batch_skip, k=self.k)
-        if x_skip is not None:
-            x = torch.cat([x, x_skip], dim=1)
-        x = self.nn(x)
-        return x, pos_skip, batch_skip
-
-
-class SegmentationModel(torch.nn.Module):
-    def __init__(self, opt,num_classes):
-        super(SegmentationModel, self).__init__()
-        self.sa1_module = SAModule(0.2, 0.2, MLP([3, 64, 64, 128]))
-        self.sa2_module = SAModule(0.25, 0.4, MLP([128 + 3, 128, 128, 256]))
-        self.sa3_module = GlobalSAModule(MLP([256 + 3, 256, 512, 1024]))
-
-        self.fp3_module = FPModule(1, MLP([1024 + 256, 256, 256]))
-        self.fp2_module = FPModule(3, MLP([256 + 128, 256, 128]))
-        self.fp1_module = FPModule(3, MLP([128, 128, 128, 128]))
-
+class SegmentationModel(UnetBasedModel):
+    def __init__(self, opt, num_classes):
+        self.down_conv_cls = SAModule
+        self.up_conv_cls = FPModule
+        self._name = 'POINTNET++_MODEL'
+        super(SegmentationModel, self).__init__(opt, num_classes)
+        
         self.lin1 = torch.nn.Linear(128, 128)
         self.lin2 = torch.nn.Linear(128, 128)
         self.lin3 = torch.nn.Linear(128, num_classes)
-        self._name = 'POINTNET++_MODEL'
 
     def forward(self, data):
-        sa0_out = (data.x, data.pos, data.batch)
-        sa1_out = self.sa1_module(*sa0_out)
-        sa2_out = self.sa2_module(*sa1_out)
-        sa3_out = self.sa3_module(*sa2_out)
-
-        fp3_out = self.fp3_module(*sa3_out, *sa2_out)
-        fp2_out = self.fp2_module(*fp3_out, *sa1_out)
-        x, _, _ = self.fp1_module(*fp2_out, *sa0_out)
-
+        """Standard forward"""
+        input = (data.x, data.pos, data.batch)
+        data_out = self.model(input)
+        data = (*data_out, *input)
+        x, _, _ = self.upconv(data)
+        
         x = F.relu(self.lin1(x))
         x = F.dropout(x, p=0.5, training=self.training)
         x = self.lin2(x)
         x = F.dropout(x, p=0.5, training=self.training)
         x = self.lin3(x)
-        return F.log_softmax(x, dim=-1)
+        return F.log_softmax(x, dim=-1)
