+++ pyproject.toml
@@ -14,6 +14,7 @@ torch-geometric = "^1.3"
 matplotlib = "^3.1"
 hydra-core = "^0.11.2"
 wandb = "^0.8.18"
+tqdm = "^4.40"
 
 [tool.poetry.dev-dependencies]
 pylint = "^2.4"
